+++ Source/equipment/infiniteregeneration-discounted.js
@@ -2,13 +2,13 @@ const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { removeModifier, addModifier, payHP } = require('../combatantDAO.js');
 const { needsLivingTargets } = require('../enemyDAO.js');
 
-module.exports = new EquipmentTemplate("Infinite Regeneration", "Pay @{hpCost} hp to grant an ally @{mod1Stacks} @{mod1}", "HP Cost / @{critBonus}", "Earth", needsLivingTargets(effect))
+module.exports = new EquipmentTemplate("Discounted Infinite Regeneration", "Pay @{hpCost} hp to grant an ally @{mod1Stacks} @{mod1}", "HP Cost / @{critBonus}", "Earth", needsLivingTargets(effect))
 	.setCategory("Pact")
 	.setTargetingTags({ target: "single", team: "delver" })
-	.setUpgrades("Discounted Infinite Regeneration", "Fate Sealing Infinite Regeneration")
+	.setSidegrades("Fate Sealing Infinite Regeneration")
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Regen", stacks: 3 }])
 	.setHpCost(50)
-	.setCost(200)
+	.setCost(100)
 	.setUses(10);
 
 function effect([target], user, isCrit, adventure) {